refactor: extract warnDuplicatePlugins to @hey-api/shared and apply in openapi-python
