+++ packages/openapi-ts/src/config/plugins.ts
@@ -1,6 +1,9 @@
-import { log } from '@hey-api/codegen-core';
 import type { AnyPluginName, PluginContext, PluginNames } from '@hey-api/shared';
-import { dependencyFactory, valueToObject } from '@hey-api/shared';
+import {
+  dependencyFactory,
+  valueToObject,
+  warnDuplicatePlugins,
+} from '@hey-api/shared';
 
 import { defaultPluginConfigs } from '../plugins/config';
 import type { Config, UserConfig } from './types';
@@ -147,51 +150,17 @@ export function getPlugins({
     }
   }
 
-  const seenPlugins = new Map<string, string>();
-
-  const stableStringify = (value: unknown): string =>
-    JSON.stringify(value, (_, v) => {
-      if (typeof v === 'function') {
-        return `[function:${(v as () => unknown).toString()}]`;
-      }
-      if (v && typeof v === 'object' && !Array.isArray(v)) {
-        return Object.fromEntries(
-          Object.entries(v as Record<string, unknown>).sort(([a], [b]) =>
-            a.localeCompare(b),
-          ),
-        );
-      }
-      return v;
-    });
-
-  const checkDuplicate = (name: string, config: Record<string, unknown>) => {
-    const serialized = stableStringify(config);
-    const previous = seenPlugins.get(name);
-    if (previous !== undefined && previous !== serialized) {
-      log.warn(
-        `Plugin "${name}" is configured more than once with conflicting options. Only the last occurrence will take effect.`,
-      );
-    }
-    seenPlugins.set(name, serialized);
-  };
+  warnDuplicatePlugins(definedPlugins as ReadonlyArray<never>);
 
   const userPlugins = definedPlugins
     .map((plugin) => {
       if (typeof plugin === 'string') {
-        checkDuplicate(plugin, {});
         return plugin;
       }
 
       const pluginName = plugin.name;
 
       if (pluginName) {
-        const config = Object.fromEntries(
-          Object.entries(plugin as unknown as Record<string, unknown>).filter(
-            ([key]) => key !== 'name',
-          ),
-        );
-        checkDuplicate(pluginName, config);
-
         // @ts-expect-error
         if (plugin.handler) {
           // @ts-expect-error
